requirements.txt
aiosqlite==0.16.1
asyncpg==0.21.0
click==7.1.2
dnspython==2.1.0
dynaconf==3.1.2
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==3.1
iso8601==0.1.14
loguru==0.5.3
pydantic==1.7.3
PyPika==0.44.1
pytz==2020.5
starlette==0.13.6
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.3
